# docs/swagger.yaml
basePath: /api
definitions:
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.InputChangePassword:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  models.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  models.TagRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  models.TagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  models.UserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
host: religious-mitzi-arasycorp-cc5d6628.koyeb.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for culinary review API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Culinary Review
  version: "1.0"
paths:
  /api/change-password:
    put:
      consumes:
      - application/json
      description: Change the password for the authenticated user
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Change Password Data
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.InputChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Change user password
      tags:
      - users
  /api/detail-user:
    get:
      consumes:
      - application/json
      description: Get user information by user ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get user by ID
      tags:
      - users
...
  /api/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and get a JWT token
      parameters:
      - description: Login Data
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - users
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - users
...
  /api/tags:
    post:
      consumes:
      - application/json
      description: Creates a new tag with the provided name.
      parameters:
      - description: Tag data to create
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.TagResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Create a new tag
      tags:
      - tags
  /api/tags/{id}:
    delete:
      description: Deletes a tag by its ID.
      parameters:
      - description: Tag ID to delete
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Tag deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Delete a tag by ID
      tags:
      - tags
    put:
      consumes:
      - application/json
      description: Updates an existing tag based on the provided data.
      parameters:
      - description: Tag ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Updated tag data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.TagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Tag updated successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Update an existing tag
      tags:
      - tags
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"